+++ engine/core/Timer.h
@@ -11,9 +11,13 @@ class Timer {
   float TotalTime() const { return total_time_; }
 
  private:
-  float last_time_ = 0.0f;
+  // Absolute timestamps are kept in double: the steady_clock epoch is a very
+  // large value (seconds since ~boot), and a 32-bit float's ~7 significant
+  // digits cannot resolve a frame-sized delta at that magnitude. delta_time_
+  // stays float because it is always small.
+  double last_time_ = 0.0;
   float delta_time_ = 0.0f;
-  float total_time_ = 0.0f;
+  double total_time_ = 0.0;
   bool first_tick_ = true;
 };
